requirements.txt
statick
pyyaml
requests
pytest
